backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==1.10.7
sentence-transformers==2.2.2
python-docx==0.8.11
PyPDF2==3.0.1
numpy==1.24.3
scikit-learn==1.2.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.9
pandas==2.0.1
transformers==4.33.3
huggingface_hub==0.17.3
